feat: Implement admin panel and secure routes
This commit introduces the admin panel, restricting access to admin users. It also secures various admin-related actions and pages, ensuring only authorized personnel can access sensitive information and perform administrative tasks.
